Fix a bug that produced repeated characters when fixing line endings
When an artificial line break is found and replaced by a whitespace, the full ngram is copied over to the output string. In this specific case, a shorter string needs to be appended to the output after the while loop is over.

Instead of using complex math calculations to discover the correct index, use "i", which is already updated with the index of last character changed.

Update tests accordingly.