Improve performance of ngramstat generation
The current code for ngramstat generation calls `fix_line_endings` for each file, which in turn requires the character_ngram from the resource_factory. This call was not cached in memory and loaded from disk every time.

Similar to NGRAMSTAT, cache the call into memory so that further calls are much faster.