Rename `get_synonyms_from_ngrams` and move it to `ngram`
PEP8 recommends the use of short module names. Also, module names containing verbs are confusing with the functions declared therein.

We should also group our ngram-related modules into a `ngram` package, so that they are comparable with word2vec embeddings from the `nn` package.

Therefore, rename it to `finder` and move it to `ngram`.