Add support for data-dependent unit tests
Some tests depend on real data to run. They should neither be run during CI nor mix up with test models.

Use pytest fixtures to override the models folders in the resource_factory. As the order of tests is not guaranteed, change scope from `session` to `module`, so that the correct models are used on remaining tests.

Add check to resource_factory tests ensure the path is correct.