Do not create duplicate ngrams (fixes #5)
When creating variants, `create_ngramstat_dump` would not check for the pre-existence of a variant as a ngram, thus generating duplicates.

As avoiding that would be hard in the current function, refactor it so that variant generation is done while we still have a dictionary with ngrams as keys, effectively merging it with `create_corpus_ngramstat_dump` (and thus avoiding common confusion). As a side effect of the merge process, add the minimum frequency to the `ngramstat.txt` filename, thus making it a readable version of the corresponding pickle file.

Expose the word ngrams as a simple dictionary to make it compatible with character ngrams interfaces.

Adapt unit tests for new expected values.

Bump ngramstat version to force its recreation.