Fix IndexError on `create_german_acronym`
When `create_german_acronym` was provided with a string with more than two whitespaces (or with symbols that would be replaced by a whitespace), the loop would fail with an IndexError.

This commit introduces `text.remove_duplicated_whitespaces`, a helper function to run a cleaning regex wherever such cleaning is necessary.

Then, use it for proper cleaning on `create_german_acronym`.

This fixes the build, as web scraping during `corpus.find_synonyms` would lead to at least one of such malformed strings.