Adapt URL-specific code to the proposed interface
Extract corpus generation from URLs out of `ngrams_web_dump`.

Also, rename `ngrams_web_dump` to `ngrams_url_dump`, so that it's not confused with the newer implementation on `get_web_ngram_stat.py`.

Move docstrings accordingly.

This is part of a major refactoring of the web scraping package to allow different implementations.