Fix a bug with checks on last acronym character
The previous code would *not* allow expansions such as "HEPA" -> "Hepatitis A" (as described in comments) and would filter out only expansions such as "HEPA" -> "Hepatitis Ba".

Add also tests for missing case of "HEPC" -> "HepaCitis A", as the previous example with "HepaTitis A" would be filtered out by `is_possible_expansion`.