Sync standards to latest cleaning routines
It has been some time that hyphens and slashes are removed during cleaning, but they were still hanging around in our standards and thus requiring special handling. They also inflate metrics of invalid acronyms, which is not undesired.

Therefore, remove uncleaned acronyms from the standards. Also add tests to two special cases ("EKG's" and "RG?s") that are also being cleaned, so that we are even clearer on the goals of the cleaning routine.