Remove NLTK tokenizer
The NLTK tokenizer was an old experiment that led to an 8% drop in F1 an so it is currently not used in the codebase.

It is also tricky to maintain, since NLTK data is not available behind some proxies and so a manual installation may be required in a domain-restricted machine.

KISS and remove it. :)